Added a basic function for kriging with this covariance function approximation. The covariance matrix of the kriged estimates is guaranteed to be positive definite (so long as the original covariance function would have guaranteed that). It falls back to dense linear algebra for the kriged points, so if you're trying to interpolate at sufficiently many points that you couldn't hold perhaps 3-4 copies of the covariance matrix in RAM, then this might fail for you. I'm thinking about scalable extensions.
I also will shortly add a wrapper struct `Maxlikdata` to go with `Maxlikopts`, where one
can wrap measurements and spatial locations in a struct that will help them not to forget
to enumerate the points in a meaningful way. I'm gathering that for new users of this
software the most common issue by far is that they forget to enumerate points according to
a KD-Tree or whatever and then get garbage results back.