Several new important updates in this release: -- distributed parallel functionality has been removed. There's a more thoughtful version in the pipes, but this is so un-tuned that it's likely to not even given you a speedup. -- The thread-parallel functionality, on the other hand, has actually been improved quite a bit and is very beneficial now, for both assembly and factorization.
In the next release, a reasonable amount of this code will be
deprecated. A lot of the maximum likelihood functions are provided with
better micro-optimizations in https://git.sr.ht/~cgeoga/GPMaxlik.jl,
which is designed to compartmentalize the efficient Gaussian likelihood
functions from the linear algebra. In this release no functionality
other than distributed-parallel has been removed, but in future release
this code will be made more lean.