Update the deps and finally switch from ThreadPools.jl to Folds.jl. While I love ThreadPools, Folds seems much more composable (by design), and after a few applications in this area I think that's the better way to go.
Still to do, though, are two TODO comments in ./src/HODLR/constructor.jl
in the derivative leaf constructors, which are manually given a
plel=false argument because something broke in a way that I really could
not debug. Honestly I don't use that part of the code anymore and it
would take a little while to write enough tests to properly try and hit
that branch, so I'm going to release this version now and when time
permits chase that other issue down.